Cleanup. Rename files.
