+++ src/components/home/__tests__/home.reducer.test.js
@@ -1,4 +1,4 @@
-import reducer from '../../../state/home-fetch'
+import reducer from '../../../state/home'
 
 describe('home reducer', () => {
   test('should return initial state', () => {
